Refactor the server manager
It's now more parallelized and, more importantly, more readable.

The current choice of Bach is probably a mistake, because Bach series do
not let you pass results on to the next item in the series - results are
only available at the *very end*. This is not what we want, because we
need the `server.properties` fragments to be passed on so we can combine
them and write them to disk. But this is a good stopping point for now,
and the structure will remain the same even if Bach is replaced.